Floats.

+++ src/main/java/net/fexcraft/mod/lib/tmt/ModelRendererTurbo.java
@@ -272,7 +272,7 @@ public class ModelRendererTurbo extends ModelRenderer {
      * @param h the height (over the y-direction)
      * @param d the depth (over the z-direction)
      */
-    public ModelRendererTurbo addBox(float x, float y, float z, int w, int h, int d){
+    public ModelRendererTurbo addBox(float x, float y, float z, float w, float h, float d){
         addBox(x, y, z, w, h, d, 0.0F);
         return this;
     }
@@ -287,7 +287,7 @@ public class ModelRendererTurbo extends ModelRenderer {
      * @param d the depth (over the z-direction)
      * @param expansion the expansion of the box. It increases the size in each direction by that many.
      */
-    public void addBox(float x, float y, float z, int w, int h, int d, float expansion){
+    public void addBox(float x, float y, float z, float w, float h, float d, float expansion){
     	addBox(x, y, z, w, h, d, expansion, 1F);
     }
     
@@ -1662,7 +1662,7 @@ public class ModelRendererTurbo extends ModelRenderer {
         GL11.glEndList();
     }
 
-	public void addShapeBox(float x, float y, float z, int w, int h, int d, float scale, float x0, float y0, float z0, float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3, float x4, float y4, float z4, float x5, float y5, float z5, float x6, float y6, float z6, float x7, float y7, float z7){
+	public void addShapeBox(float x, float y, float z, float w, float h, float d, float scale, float x0, float y0, float z0, float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3, float x4, float y4, float z4, float x5, float y5, float z5, float x6, float y6, float z6, float x7, float y7, float z7){
 		float f4 = x + w, f5 = y + h, f6 = z + d;
 		x -= scale; y -= scale; z -= scale;
 		f4 += scale; f5 += scale; f6 += scale;
